Connect Plate to editor slice
